refactor(encoders): code style inheritance refactoring
