+++ jina/executors/encoders/nlp/flair.py
@@ -5,11 +5,11 @@ from typing import Union, Tuple, List
 
 import numpy as np
 
-from ..frameworks import BaseTextTorchEncoder
+from ..frameworks import BaseTorchEncoder
 from ...decorators import batching, as_ndarray
 
 
-class FlairTextEncoder(BaseTextTorchEncoder):
+class FlairTextEncoder(BaseTorchEncoder):
     """
     :class:`FlairTextEncoder` encodes data from an array of string in size `B` into a ndarray in size `B x D`.
     Internally, :class:`FlairTextEncoder` wraps the DocumentPoolEmbeddings from Flair.
