+++ tests/unit/executors/encoders/video/test_paddlehub.py
@@ -1,4 +1,4 @@
-from jina.executors.encoders.video.paddlehub import VideoPaddlehubEncoder
+from jina.executors.encoders.video.paddlehub import VideoPaddleEncoder
 from tests.unit.executors.encoders.video import VideoTestCase
 
 
@@ -6,4 +6,4 @@ class VideoPaddleHubTestCase(VideoTestCase):
     def _get_encoder(self, metas):
         self.target_output_dim = 2048
         self.input_dim = 224
-        return VideoPaddlehubEncoder(metas=metas)
+        return VideoPaddleEncoder(metas=metas)
